requirements.txt
asgiref==3.4.1
autopep8==1.6.0
bleach==4.1.0
certifi==2021.10.8
charset-normalizer==2.0.9
cssselect2==0.4.1
Django==3.2.9
django-appconf==1.0.5
django-autoslug==1.9.8
django-classy-tags==2.0.0
django-cms==3.9.0
django-filer==2.1.1
-e git+https://github.com/tom-vndm/django-fobi@d3488acc1e01b592b3783b58bb07a9988e3fe7aa#egg=django_fobi
django-formtools==2.3
django-imagekit==4.1.0
django-js-asset==1.2.2
django-modeltranslation==0.17.3
django-mptt==0.13.4
django-nine==0.2.5
django-nonefield==0.4
django-polymorphic==3.0.0
django-sekizai==2.0.0
django-treebeard==4.5.1
djangocms-admin-style==2.0.2
djangocms-attributes-field==2.0.0
djangocms-file==3.0.0
djangocms-googlemap==2.0.0
djangocms-link==3.0.0
djangocms-picture==3.0.0
djangocms-snippet==3.0.0
djangocms-style==3.0.0
djangocms-text-ckeditor==4.0.0
djangocms-video==3.0.0
easy-thumbnails==2.8
html5lib==1.1
idna==3.3
lxml==4.6.4
packaging==21.3
pilkit==2.0
Pillow==8.4.0
pycodestyle==2.8.0
pyparsing==3.0.6
pytz==2021.3
reportlab==3.6.2
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
svglib==1.1.0
tinycss2==1.1.0
toml==0.10.2
Unidecode==1.1.2
urllib3==1.26.7
vishap==0.1.5
webencodings==0.5.1
